major refactoring of ClearSigning related code

+++ packages/ethereum-contracts/tasks/deploy-permit2-clearsigning-macro-forwarder.sh
@@ -3,17 +3,18 @@ set -eu
 set -o pipefail
 
 # Usage:
-# tasks/deploy-712-macro-forwarder.sh <network>
+# tasks/deploy-permit2-clearsigning-macro-forwarder.sh <network>
 #
 # The invoking account needs to be (co-)owner of the resolver and governance
 #
 # important ENV vars:
-# RELEASE_VERSION, ONLY712MACROFWD_DEPLOYER_PK
+# RELEASE_VERSION, PERMIT2CLEARSIGNINGMACROFWD_DEPLOYER_PK
+# Fallback for backwards compatibility: PERMIT2MACROFWD_DEPLOYER_PK
 # EXPECTED_ADDRESS: set after generating deployer with vanity-eth
-#   (e.g. npx vanityeth -i 712f --contract), or set SKIP_ADDRESS_CHECK=1 to skip.
+#   (e.g. npx vanityeth -i 2712 --contract), or set SKIP_ADDRESS_CHECK=1 to skip.
 #
 # You can use the npm package vanity-eth to get a deployer account for a given contract address:
-# Example use: npx vanityeth -i 712f --contract
+# Example use: npx vanityeth -i 2712 --contract
 #
 # For optimism the gas estimation doesn't work, requires setting EST_TX_COST
 # (the value auto-detected for arbitrum should work).
@@ -26,15 +27,22 @@ source .env
 set -x
 
 network=$1
-expectedContractAddr=${EXPECTED_ADDRESS:-"0x712Fc5863F53AFBa980207006cfd74F6c25fE055"}
-deployerPk=$ONLY712MACROFWD_DEPLOYER_PK
+expectedContractAddr=${EXPECTED_ADDRESS:-"0x712F1ccD0472025EC75bB67A92AA6406cDA0031D"}
+
+deployerPk=${PERMIT2CLEARSIGNINGMACROFWD_DEPLOYER_PK:-${PERMIT2MACROFWD_DEPLOYER_PK:-}}
+
+skipDeploy=${SKIP_DEPLOY:-0}
 
 tmpfile="/tmp/$(basename "$0").addr"
 
 # deploy
-DETERMINISTIC_DEPLOYER_PK=$deployerPk npx truffle exec --network "$network" ops-scripts/deploy-deterministically.js : Only712MacroForwarder | tee "$tmpfile"
-contractAddr=$(tail -n 1 "$tmpfile")
-rm "$tmpfile"
+if [ "$skipDeploy" -eq 0 ]; then
+    DETERMINISTIC_DEPLOYER_PK=$deployerPk npx truffle exec --network "$network" ops-scripts/deploy-deterministically.js : Permit2ClearSigningMacroForwarder | tee "$tmpfile"
+    contractAddr=$(tail -n 1 "$tmpfile")
+    rm "$tmpfile"
+else
+    contractAddr=$expectedContractAddr
+fi
 
 echo "deployed to $contractAddr"
 if [[ -n "$expectedContractAddr" && $contractAddr != "$expectedContractAddr" ]]; then
@@ -48,11 +56,11 @@ fi
 sleep 5
 # allow to fail
 set +e
-npx truffle run --network "$network" verify Only712MacroForwarder@"$contractAddr"
+npx truffle run --network "$network" verify Permit2ClearSigningMacroForwarder@"$contractAddr"
 set -e
 
 # set resolver
-ALLOW_UPDATE=1 npx truffle exec --network "$network" ops-scripts/resolver-set-key-value.js : Only712MacroForwarder "$contractAddr"
+ALLOW_UPDATE=1 npx truffle exec --network "$network" ops-scripts/resolver-set-key-value.js : Permit2ClearSigningMacroForwarder "$contractAddr"
 
 # create gov action
 npx truffle exec --network "$network" ops-scripts/gov-set-trusted-forwarder.js : 0x0000000000000000000000000000000000000000 "$contractAddr" 1
